DRY/consolidation/reduce work on providers
The efforts in this PR are to try and consolidate and unify the provider
classes to have a single base class. This has meant multiple changes to
existing files and some new methodology.

* credential_form has been moved to the common/provider.py file and has
  been augmented with all cloud and infra credential types
* Infra and Cloud now have their own base classes which set up certain
  base attributes to assist in making the right decisions when it comes to
  string naming, quadicon matching and forms.
* create, update, delete, delete_if_exists, validate, load_all_*
  refresh_provider_relationships, get_detail, _do_stats_match,
  num_template, num_vm, exists, wait_for_delete and all policy methods
  have been removed from both the Infra/Cloud provider class and have been
  moved into the common class.
* SCVMM custom fill method has been removed and functionality for
  handling the special domain has been incorporated into the other fill
  functions.
* get_from_config and get_credentials_from_config no longer exist in the
  provider modules and are now called get_crud and
  get_credentials_from_config and live in the utils.providers
  module. These are special in that they now are no longer class
  specific. get_from_config will return an Infra or Cloud object,
  depending on the key of the provider loaded.
* Big change in how credentials are passed into the provider
  objects. Previously there was a credentials attribute which stored the
  "default" credentials and then other attributes that stored credentials
  like amqp and candu. This prevented the unification of the provider
  classes. To combat this, the providers credentials attribute is now a
  dict. The credentials key name is actually not important as the
  credential object still has the attributes that denote it's type, ie
  candu and the forms are handled accordingly. As a result, all tests that
  "add" or mess with credentials will need to be aware of this change. The
  "update" methods seem to work fine with the dict credentials
  approach. This _has_ been tested, but if you find any issues, please let
  us know.
* utils.providers.provider_factory function has been renamed get_mgmt to
  provider better specificity and bring it inline with the other methods.
