+++ cfme/infrastructure/cluster.py
@@ -7,10 +7,10 @@
 
 import ui_navigate as nav
 from cfme.fixtures import pytest_selenium as sel
-from cfme.infrastructure import provider
 from cfme.web_ui import Quadicon, Region, listaccordion as list_acc, paginator, toolbar as tb
 from functools import partial
 from utils.pretty import Pretty
+from utils.providers import get_crud
 from utils.wait import wait_for
 
 details_page = Region(infoblock_type='detail')
@@ -50,7 +50,7 @@ class Cluster(Pretty):
     def __init__(self, name=None, provider_key=None):
         self.name = name
         if provider_key:
-            self.provider = provider.get_from_config(provider_key)
+            self.provider = get_crud(provider_key)
         else:
             self.provider = None
