+++ cfme/tests/infrastructure/test_esx_direct_host.py
@@ -6,9 +6,10 @@ not be difficult to extend the parametrizer.
 import pytest
 import random
 
-from cfme.infrastructure.provider import VMwareProvider, get_from_config
+from cfme.infrastructure.provider import VMwareProvider
 from utils.conf import cfme_data, credentials
 from utils.net import resolve_hostname
+from utils.providers import get_crud
 from utils.version import Version
 from utils.wait import wait_for
 
@@ -56,7 +57,7 @@ def pytest_generate_tests(metafunc):
             name=host["name"],
             hostname=host["name"],
             ip_address=ip_address,
-            credentials=cred,
+            credentials={'default': cred},
             provider_data=provider_data,
         )
         arg_values.append([host_provider, provider_data, provider_key])
@@ -66,7 +67,7 @@ def pytest_generate_tests(metafunc):
 
 @pytest.yield_fixture(scope="module")
 def setup_provider(provider, original_provider_key):
-    original_provider = get_from_config(original_provider_key)
+    original_provider = get_crud(original_provider_key)
     if original_provider.exists:
         # Delete original provider's hosts first
         for host in original_provider.hosts:
