+++ cfme/tests/test_rest.py
@@ -6,7 +6,7 @@ import pytest
 from cfme.configure.configuration import server_roles_disabled
 
 from utils import error, mgmt_system, testgen
-from utils.providers import setup_a_provider as _setup_a_provider, provider_factory
+from utils.providers import setup_a_provider as _setup_a_provider, get_mgmt
 from utils.version import current_version
 from utils.virtual_machines import deploy_template
 from utils.wait import wait_for
@@ -181,7 +181,7 @@ def test_provider_refresh(request, setup_a_provider, rest_api):
     """
     if "refresh" not in rest_api.collections.providers.action.all:
         pytest.skip("Refresh action is not implemented in this version")
-    provider_mgmt = provider_factory(setup_a_provider.key)
+    provider_mgmt = get_mgmt(setup_a_provider.key)
     provider = rest_api.collections.providers.find_by(name=setup_a_provider.name)[0]
     with server_roles_disabled("ems_inventory", "ems_operations"):
         vm_name = deploy_template(
@@ -274,7 +274,7 @@ def test_provider_crud(request, rest_api, from_detail):
 def vm(request, setup_a_provider, rest_api):
     if "refresh" not in rest_api.collections.providers.action.all:
         pytest.skip("Refresh action is not implemented in this version")
-    provider_mgmt = provider_factory(setup_a_provider.key)
+    provider_mgmt = get_mgmt(setup_a_provider.key)
     provider = rest_api.collections.providers.find_by(name=setup_a_provider.name)[0]
     vm_name = deploy_template(
         setup_a_provider.key,
